seperated logger into another module
